+++ packages/create-theme/public/template/docs/.vitepress/config.ts
@@ -77,7 +77,7 @@ export default defineConfig({
     logo: '/logo.png',
     // editLink: {
     //   pattern:
-    //     'https://github.com/ATQQ/sugar-blog/tree/master/packages/blogpress/:path',
+    //     'https://github.com/xfy196/blog/tree/master/packages/blogpress/:path',
     //   text: '去 GitHub 上编辑内容'
     // },
     nav: [
@@ -87,7 +87,7 @@ export default defineConfig({
     socialLinks: [
       {
         icon: 'github',
-        link: 'https://github.com/ATQQ/sugar-blog/tree/master/packages/theme'
+        link: 'https://github.com/xfy196/blog/tree/master/packages/theme'
       }
     ]
   }