+++ packages/theme/docs/.vitepress/config.ts
@@ -65,12 +65,12 @@ export default defineConfig({
     socialLinks: [
       {
         icon: 'github',
-        link: 'https://github.com/ATQQ/sugar-blog/tree/master/packages/theme'
+        link: 'https://github.com/xfy196/blog/tree/master/packages/theme'
       }
     ],
     editLink: {
       pattern:
-        'https://github.com/ATQQ/sugar-blog/tree/master/packages/theme/docs/:path',
+        'https://github.com/xfy196/blog/tree/master/packages/theme/docs/:path',
       text: '去 GitHub 上编辑内容'
     },
     lastUpdatedText: '上次更新于'