+++ packages/theme/docs/changelog.md
@@ -49,7 +49,7 @@ pnpm up 小小荧
 
 ### Patch Changes
 
-- feat: 新的项目创建方式使用 [@sugarat/create-theme](https://github.com/ATQQ/sugar-blog/tree/master/packages/create-theme) 快速创建模板项目
+- feat: 新的项目创建方式使用 [@sugarat/create-theme](https://github.com/xfy196/blog/tree/master/packages/create-theme) 快速创建模板项目
 - feat: 支持使用自定义的 HTML 内容设置文章在首页卡片列表里展示的 `描述信息`
 - feat: 支持使用 `cover:false` 隐藏首页展示的封面
 
@@ -156,7 +156,7 @@ pnpm up 小小荧
 - feat: 内置 pagefind 搜索针对中文优化
 - fix: 内置 pagefind 在自定义 base 后无效
 - chore: 没有 git 仓库时，打印 Not 日志
-- chore: pagefind 引导使用独立的插件 [vitepress-plugin-pagefind](https://github.com/ATQQ/sugar-blog/blob/master/packages/vitepress-plugin-pagefind/README-zh.md)
+- chore: pagefind 引导使用独立的插件 [vitepress-plugin-pagefind](https://github.com/xfy196/blog/blob/master/packages/vitepress-plugin-pagefind/README-zh.md)
 
 ## 0.1.19 (2023/04/02)
 
@@ -183,7 +183,7 @@ pnpm up 小小荧
 
 ## 0.1.15 (2023/03/26)
 
-- feat: 分离可独立使用的的 搜索插件[pagefind](https://github.com/ATQQ/sugar-blog/tree/master/packages/vitepress-plugin-pagefind)
+- feat: 分离可独立使用的的 搜索插件[pagefind](https://github.com/xfy196/blog/tree/master/packages/vitepress-plugin-pagefind)
 - feat: 搜索框展示 UI 优化，支持定制文案，展示适配移动端
 - fix: 搜索结果未全部展示
 - fix: 中文文件名文章时间显示 NaN