+++ packages/vitepress-plugin-pagefind/CHANGELOG.md
@@ -59,4 +59,4 @@
 
 ### Minor Changes
 
-- feat: 从 [小小荧](https://github.com/ATQQ/sugar-blog/tree/master/packages/theme) 主题分离独立的 pagefind 插件
+- feat: 从 [小小荧](https://github.com/xfy196/blog/tree/master/packages/theme) 主题分离独立的 pagefind 插件